+++ src/idc/reader/__init__.py
@@ -2,3 +2,4 @@ from ._data import DataReader
 from ._multi import MultiReader
 from ._poll_dir import PollDir
 from ._pyfunc import PythonFunctionReader
+from ._watch_dir import WatchDir, EVENTS, EVENT_MODIFIED, EVENT_CREATED