+++ src/idc/filter/_load_data.py
@@ -1,0 +1,114 @@
+import argparse
+from typing import List
+
+from wai.logging import LOGGING_WARNING
+
+from idc.api import DATATYPES, data_type_to_class, DataTypeSupporter, ImageData
+from kasperl.api import make_list, flatten_list
+from seppl.io import BatchFilter
+
+
+class LoadData(BatchFilter, DataTypeSupporter):
+    """
+    Removes the specified classes/labels.
+    """
+
+    def __init__(self, data_type: str = None,
+                 logger_name: str = None, logging_level: str = LOGGING_WARNING):
+        """
+        Initializes the filter.
+
+        :param data_type: the type of output to generate from the images
+        :type data_type: str
+        :param logger_name: the name to use for the logger
+        :type logger_name: str
+        :param logging_level: the logging level to use
+        :type logging_level: str
+        """
+        super().__init__(logger_name=logger_name, logging_level=logging_level)
+        self.data_type = data_type
+        self._output_cls = None
+
+    def name(self) -> str:
+        """
+        Returns the name of the handler, used as sub-command.
+
+        :return: the name
+        :rtype: str
+        """
+        return "load-data"
+
+    def description(self) -> str:
+        """
+        Returns a description of the handler.
+
+        :return: the description
+        :rtype: str
+        """
+        return "Wraps the incoming file name in the specified data type and forwards it. The input file name can contain placeholders."
+
+    def accepts(self) -> List:
+        """
+        Returns the list of classes that are accepted.
+
+        :return: the list of classes
+        :rtype: list
+        """
+        return [str]
+
+    def generates(self) -> List:
+        """
+        Returns the list of classes that get produced.
+
+        :return: the list of classes
+        :rtype: list
+        """
+        if self.data_type is None:
+            return [ImageData]
+        else:
+            return [data_type_to_class(self.data_type)]
+
+    def _create_argparser(self) -> argparse.ArgumentParser:
+        """
+        Creates an argument parser. Derived classes need to fill in the options.
+
+        :return: the parser
+        :rtype: argparse.ArgumentParser
+        """
+        parser = super()._create_argparser()
+        parser.add_argument("-t", "--data_type", choices=DATATYPES, type=str, default=None, help="The type of data to forward", required=True)
+        return parser
+
+    def _apply_args(self, ns: argparse.Namespace):
+        """
+        Initializes the object with the arguments of the parsed namespace.
+
+        :param ns: the parsed arguments
+        :type ns: argparse.Namespace
+        """
+        super()._apply_args(ns)
+        self.data_type = ns.data_type
+
+    def initialize(self):
+        """
+        Initializes the processing, e.g., for opening files or databases.
+        """
+        super().initialize()
+        if self.data_type is None:
+            raise Exception("No data type defined!")
+        self._output_cls = data_type_to_class(self.data_type)
+
+    def _do_process(self, data):
+        """
+        Processes the data record(s).
+
+        :param data: the record(s) to process
+        :return: the potentially updated record(s)
+        """
+        result = []
+
+        for item in make_list(data):
+            item_new = self._output_cls(source=self.session.expand_placeholders(item))
+            result.append(item_new)
+
+        return flatten_list(result)