+++ CHANGES.rst
@@ -43,6 +43,7 @@ Changelog
 - `od-to-is` now handles labels or regexp pattern correctly
 - added `is-to-od` filter that generates object detection annotations from contours determined in image segmentation layers
 - added `to-metadata` writer that outputs the meta-data of an image
+- added `attach-metadata` filter that loads meta-data from a directory and attaches it to the data passing through