+++ CHANGES.rst
@@ -45,6 +45,8 @@ Changelog
 - added `to-metadata` writer that outputs the meta-data of an image
 - added `attach-metadata` filter that loads meta-data from a directory and attaches it to the data passing through
 - added `load-data` filter to turn file names into data containers
+- added `annotation-to-storage` and `annotation-from-storage` filters
+- annotation data is now being type-checked when setting it